[refactor] 가격범위 전달객체 디렉토리 위치변경 (KakaoFunding#183)
