--- src/main/java/org/kakaoshare/backend/domain/rank/dto/PriceRange.java
@@ -1,11 +1,0 @@
-package org.kakaoshare.backend.domain.rank.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class PriceRange {
-    private int minPrice;
-    private int maxPrice;
-}
