+++ src/main/java/org/kakaoshare/backend/domain/rank/util/RankStrategy.java
@@ -1,7 +1,7 @@
 package org.kakaoshare.backend.domain.rank.util;
 
 import java.util.List;
-import org.kakaoshare.backend.domain.rank.dto.PriceRange;
+import org.kakaoshare.backend.common.vo.PriceRange;
 import org.kakaoshare.backend.domain.rank.dto.RankResponse;
 
 public interface RankStrategy {
